+++ tests/unit/common/test_osclients.py
@@ -904,23 +904,6 @@ class OSClientsTestCase(test.TestCase):
                 key += "%s" % {"version": version}
             self.assertEqual(fake_designate, self.clients.cache[key])
 
-    def test_senlin(self):
-        mock_senlin = mock.MagicMock()
-        self.assertNotIn("senlin", self.clients.cache)
-        with mock.patch.dict("sys.modules", {"senlinclient": mock_senlin}):
-            client = self.clients.senlin()
-            self.assertEqual(mock_senlin.client.Client.return_value, client)
-            mock_senlin.client.Client.assert_called_once_with(
-                "1",
-                username=self.credential.username,
-                password=self.credential.password,
-                project_name=self.credential.tenant_name,
-                cert=self.credential.cacert,
-                auth_url=self.credential.auth_url)
-            self.assertEqual(
-                mock_senlin.client.Client.return_value,
-                self.clients.cache["senlin"])
-
     @mock.patch("%s.Magnum._get_endpoint" % PATH)
     def test_magnum(self, mock_magnum__get_endpoint):
         fake_magnum = fakes.FakeMagnumClient()